requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
qrcode>=7.4.0
pillow>=10.0.0